requirements.txt
gym
spriteworld
tqdm
torch
wandb
hydra-core
omegaconf
numpy
pillow
moviepy
imageio
h5py
coverage
black
stable-baselines3
imageio-ffmpeg
moviepy
imageio
pyvirtualdisplay
portpicker
gym_unity
pandas
tensorboard
causal_world
timm==0.3.2
